% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sormasr.R
\name{Sormasr}
\alias{Sormasr}
\title{Connect and pull/get data from a SORMAS}
\format{
An R6 class called Sormasr.
}
\description{
To create a SORMAS connection, you need a SORMAS base URL, username, password, and an API version
The R6 Class called `Sormasr` representing a SORMAS instance connection
}
\details{
You can use a SORMAS connection to get data several times without needing to manually supply your user credentials on each API call.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{request_sent}}{The request used to perform an API call}

\item{\code{name}}{Name of the user}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Sormasr-new}{\code{Sormasr$new()}}
\item \href{#method-Sormasr-get_states}{\code{Sormasr$get_states()}}
\item \href{#method-Sormasr-get_disease_data}{\code{Sormasr$get_disease_data()}}
\item \href{#method-Sormasr-clone}{\code{Sormasr$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sormasr-new"></a>}}
\if{latex}{\out{\hypertarget{method-Sormasr-new}{}}}
\subsection{Method \code{new()}}{
Create a connection to SORMAS using basic authentication
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sormasr$new(base_url, username, password, api_version = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{Base url e.g https://}

\item{\code{username}}{Registered username e.g}

\item{\code{password}}{Registered password e.g}

\item{\code{api_version}}{The api version e.g}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Sormasr` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sormasr-get_states"></a>}}
\if{latex}{\out{\hypertarget{method-Sormasr-get_states}{}}}
\subsection{Method \code{get_states()}}{
Get states info from SORMAS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sormasr$get_states(page, size = 37)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{page}}{number of page zero indexed i.e starts from 0}

\item{\code{size}}{size starts from 1 to 37}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sormasr-get_disease_data"></a>}}
\if{latex}{\out{\hypertarget{method-Sormasr-get_disease_data}{}}}
\subsection{Method \code{get_disease_data()}}{
Get disease from SORMAS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sormasr$get_disease_data(
  disease = c("COVID", "CSM", "Cholera", "Lassa", "Measles", "MonkeyPox", "YellowFever"),
  stateId,
  week,
  year
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{disease}}{vector of disease from "COVID", "CSM", "Cholera", "Lassa", "Measles", "MonkeyPox", "YellowFever"}

\item{\code{stateId}}{vector of stateId}

\item{\code{week}}{vector of}

\item{\code{year}}{vector of}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of all possible fields for a specific metadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sormasr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Sormasr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sormasr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
